libticalcs: widen key codes argument of ticalcs_calc_send_key() and CalcFncts.send_key() from 16 to 32 bits, so that ticalcs_calc_send_key() can be used for Nspire keypresses.

+++ libticalcs/trunk/src/calc_00.c
@@ -38,7 +38,7 @@ int noop_is_ready (CalcHandle* handle)
 	return 0;
 }
 
-int noop_send_key (CalcHandle* handle, uint16_t key)
+int noop_send_key (CalcHandle* handle, uint32_t key)
 {
 	(void)handle, (void)key;
 	return 0;
